+++ frontend/lost/src/containers/pipeline/start/2/nodes/DataExportNode.jsx
@@ -1,14 +1,13 @@
-import React from 'react'
 import { faCloudDownloadAlt } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import VerificationTitle from './VerificationTitle'
-import NodeBody from './NodeBody'
+import NodeBody from '../../../globalComponents/node-structure/NodeBody'
+import NodeHeader from '../../../globalComponents/node-structure/NodeHeader'
 
 const DataExportNode = (props) => {
     return (
         <div className="graph-node">
-            <VerificationTitle
-                verified={props.verified}
+            <NodeHeader
+                bgColorClass={props.verified ? 'bg-green' : 'bg-orange'}
                 title={props.title}
                 icon={faCloudDownloadAlt}
             />